refactor: Extract util/shell from util/system
This is a minimal extraction of shell-specific functions
from util/system.
