+++ src/Makefile.am
@@ -291,6 +291,7 @@ BITCOIN_CORE_H = \
   util/result.h \
   util/serfloat.h \
   util/settings.h \
+  util/shell.h \
   util/sock.h \
   util/spanparsing.h \
   util/string.h \
@@ -707,6 +708,7 @@ libbitcoin_util_a_SOURCES = \
   util/threadinterrupt.cpp \
   util/threadnames.cpp \
   util/serfloat.cpp \
+  util/shell.cpp \
   util/spanparsing.cpp \
   util/strencodings.cpp \
   util/string.cpp \
@@ -949,6 +951,7 @@ libbitcoinkernel_la_SOURCES = \
   util/rbf.cpp \
   util/serfloat.cpp \
   util/settings.cpp \
+  util/shell.cpp \
   util/strencodings.cpp \
   util/string.cpp \
   util/syscall_sandbox.cpp \
