+++ src/test/fuzz/util.h
@@ -26,6 +26,7 @@
 #include <array>
 #include <cstdint>
 #include <cstdio>
+#include <memory>
 #include <optional>
 #include <string>
 #include <vector>
