+++ src/test/run_command_tests.cpp
@@ -21,7 +21,7 @@
 
 #include <boost/test/unit_test.hpp>
 
-BOOST_FIXTURE_TEST_SUITE(system_tests, BasicTestingSetup)
+BOOST_FIXTURE_TEST_SUITE(run_command_tests, BasicTestingSetup)
 
 // At least one test is required (in case ENABLE_EXTERNAL_SIGNER is not defined).
 // Workaround for https://github.com/bitcoin/bitcoin/issues/19128
