+++ src/util/shell.cpp
@@ -1,0 +1,64 @@
+// Copyright (c) 2009-2010 Satoshi Nakamoto
+// Copyright (c) 2009-2022 The Bitcoin Core developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#include <util/shell.h>
+
+#if defined(HAVE_CONFIG_H)
+#include <config/bitcoin-config.h>
+#endif
+
+#ifdef ENABLE_EXTERNAL_SIGNER
+#if defined(__GNUC__)
+// Boost 1.78 requires the following workaround.
+// See: https://github.com/boostorg/process/issues/235
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wnarrowing"
+#endif
+#include <boost/process.hpp>
+#include <memory>
+#include <ostream>
+#include <vector>
+#if defined(__GNUC__)
+#pragma GCC diagnostic pop
+#endif
+#endif // ENABLE_EXTERNAL_SIGNER
+
+#include <logging.h>
+#include <tinyformat.h>
+#include <univalue.h>
+#include <util/string.h>
+
+#include <cstdlib>
+#include <stdexcept>
+#include <string>
+
+#ifdef WIN32
+#include <compat/compat.h>
+#include <codecvt>
+#include <locale>
+#endif // WIN32
+
+#ifndef WIN32
+std::string ShellEscape(const std::string& arg)
+{
+    std::string escaped = arg;
+    ReplaceAll(escaped, "'", "'\"'\"'");
+    return "'" + escaped + "'";
+}
+#endif
+
+#if HAVE_SYSTEM
+void runCommand(const std::string& strCommand)
+{
+    if (strCommand.empty()) return;
+#ifndef WIN32
+    int nErr = ::system(strCommand.c_str());
+#else
+    int nErr = ::_wsystem(std::wstring_convert<std::codecvt_utf8_utf16<wchar_t>,wchar_t>().from_bytes(strCommand).c_str());
+#endif
+    if (nErr)
+        LogPrintf("runCommand error: system(%s) returned %d\n", strCommand, nErr);
+}
+#endif
